remove auto reset when no USB port detected
The implementation of this feature was not reliable since it relied
on the host sending USB start of frame packets.  On windows, if
no software communicated with the device for a second or two, it
would stop sending start of frame packets and the device would
reset by itself.

A more accurate way would be to check that any USB packet has been sent,
and only calling `wdt_reset()` to prevent the reset. Will need to reduce
the flash size to get this to work though, so leave it unimplemented
for now.